pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graphsmith"
version = "0.2.0"
description = "Typed contract-based subgraph standard with deterministic runtime, local registry, and LLM planner"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Typing :: Typed",
]
dependencies = [
  "pydantic>=2.6",
  "typer>=0.12",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
llm = [
  "httpx>=0.27",
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "httpx>=0.27",
]

[project.scripts]
graphsmith = "graphsmith.cli.main:app"

[tool.setuptools.packages.find]
include = ["graphsmith*"]

[tool.setuptools.package-data]
graphsmith = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
